host  관련 삭제
Host를 받고있던 관련 entity를 수정
Host Entity 삭제
