+++ src/main/java/com/weble/linkedhouse/customer/controller/HostController.java
@@ -1,0 +1,4 @@
+package com.weble.linkedhouse.customer.controller;
+
+public class HostController {
+}
