--- src/main/java/com/weble/linkedhouse/host/entity/constant/ApproveState.java
@@ -1,5 +1,0 @@
-package com.weble.linkedhouse.host.entity.constant;
-
-public enum ApproveState {
-    APPROVE, NOT_APPROVE
-}
