--- src/main/java/com/weble/linkedhouse/host/entity/constant/LockState.java
@@ -1,5 +1,0 @@
-package com.weble.linkedhouse.host.entity.constant;
-
-public enum LockState {
-    STOP, LIVE
-}
