+++ src/main/java/com/weble/linkedhouse/review/entity/FeedbackHost.java
@@ -1,7 +1,6 @@
 package com.weble.linkedhouse.review.entity;
 
 import com.weble.linkedhouse.customer.entity.Customer;
-import com.weble.linkedhouse.host.entity.Host;
 import com.weble.linkedhouse.util.AuditingFields;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -9,7 +8,6 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
@@ -26,12 +24,10 @@ public class FeedbackHost extends AuditingFields {
     private Long feedbackHostId;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "customerId")
     private Customer customer;      //리뷰 대상자
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "hostId")
-    private Host host;      // 리뷰쓰는 사람
+    private Customer writer;      // 리뷰쓰는 사람
 
     private String content;
 
@@ -40,15 +36,15 @@ public class FeedbackHost extends AuditingFields {
     @Column(name = "damage_degree")
     private int damageDegree;  //파손
 
-    private FeedbackHost(Customer customer, Host host, String content, int attitude, int damageDegree) {
+    private FeedbackHost(Customer customer, Customer writer, String content, int attitude, int damageDegree) {
         this.customer = customer;
-        this.host = host;
+        this.writer = writer;
         this.content = content;
         this.attitude = attitude;
         this.damageDegree = damageDegree;
     }
 
-    public static FeedbackHost of(Customer customer, Host host, String content, int attitude, int damageDegree) {
-        return new FeedbackHost(customer, host, content, attitude, damageDegree);
+    public static FeedbackHost of(Customer customer, Customer writer, String content, int attitude, int damageDegree) {
+        return new FeedbackHost(customer, writer, content, attitude, damageDegree);
     }
 }
